added Lottie animation
