+++ App.js
@@ -1,25 +1,27 @@
 import React from 'react';
-import {StyleSheet, View, Text} from 'react-native'
+import { StyleSheet, View } from 'react-native'
 
-import Header from './components/Header';
+import Start from './components/Start';
 import ActivityApi from './components/ActivityApi';
 
 
 const styles = StyleSheet.create({
 	container: {
 		flex: 1,
-		backgroundColor: '#FAF6F6',
+		backgroundColor: '#C5DECD',
 		flexDirection: 'column',
 		justifyContent: 'center',
 		alignItems: 'center'
-	}, 
+	},
 })
 
+
 const App = () => {
+
 	return (
 		<View style={styles.container}>
-			<Header/>
-			<ActivityApi/>
+			<Start />
+			<ActivityApi />
 		</View>
 	);
 };
