+++ package.json
@@ -17,7 +17,8 @@
     "react-dom": "17.0.1",
     "react-native": "0.64.2",
     "react-native-web": "0.17.1",
-    "styled-components": "^5.3.3"
+    "styled-components": "^5.3.3",
+    "lottie-react-native": "4.0.3"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
